backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
werkzeug==3.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0